Payara 5 merge in current master (#931)
Merge current master into Payara-5 branch

+++ appserver/jdbc/jdbc-ra/jdbc-core/src/main/java/com/sun/gjc/spi/ManagedConnectionFactoryImpl.java
@@ -1160,12 +1160,12 @@ public abstract class ManagedConnectionFactoryImpl implements javax.resource.spi
     
     public void setSlowQueryThresholdInSeconds(String seconds) {
         spec.setDetail(DataSourceSpec.SLOWSQLLOGTHRESHOLD, seconds);
-        int threshold = Integer.valueOf(seconds);
+        double threshold = Double.valueOf(seconds);
         if (threshold > 0) {
             if (sqlTraceDelegator == null) {
                 sqlTraceDelegator = new SQLTraceDelegator(getPoolName(), getApplicationName(), getModuleName());
             }
-            sqlTraceDelegator.registerSQLTraceListener(new SlowSQLLogger(threshold*1000));
+            sqlTraceDelegator.registerSQLTraceListener(new SlowSQLLogger((long) (threshold*1000)));
         }
     }
